Add support for dynamic attributes in Carrier
Carrier attributes may be dynamic, with these values calculated by
Exporter prior to being used in ETEngine.
